remove token metadata deps from utils and update it a bit
