+++ core/rust/utils/src/lib.rs
@@ -1,6 +1,8 @@
+pub use account::*;
 pub use assertions::*;
 pub use misc::*;
 
+mod account;
 mod assertions;
 mod misc;
